Put helper files into separate projects and todos helpers
