+++ src/controllers/helpers/todos/todoHelpers.js
@@ -2,7 +2,7 @@ import {
   showTodoBody
 } from './todoBody';
 
-import todosController from "../todosController";
+import todosController from "../../todosController";
 
 import {
   extractProject,
